directory clean:./ → src
Change to a directory structure with the src folder as the main axis
