+++ src/components/ListDetail.tsx
