+++ src/components/contact.tsx
@@ -1,10 +1,10 @@
 /** @jsxImportSource theme-ui */
 
-import Divider from "../elements/divider"
-import Inner from "../elements/Inner"
-import Content from "../elements/content"
+import Divider from "elements/divider"
+import Inner from "elements/Inner"
+import Content from "elements/content"
 import Svg from "./Svg"
-import { UpDown, UpDownWide, waveAnimation } from "../style/animation"
+import { UpDown, UpDownWide, waveAnimation } from "style/animation"
 import Footer from "./Footer"
 import ContactMDX from "../sections/contact.mdx"
