+++ src/pages/users/index.tsx
@@ -1,10 +1,10 @@
 import { GetStaticProps } from 'next'
 import Link from 'next/link'
 
-import { User } from '../../interfaces'
-import { sampleUserData } from '../../utils/sample-data'
-import Layout from '../../components/Layout'
-import List from '../../components/List'
+import { User } from '../../../interfaces'
+import { sampleUserData } from '../../../utils/sample-data'
+import Layout from '../../../components/Layout'
+import List from '../../../components/List'
 
 type Props = {
   items: User[]
