+++ src/sections/projects.mdx
