+++ src/sections/structured.mdx
