+++ src/sections/study.mdx
