+++ src/style/animation.tsx
