+++ tsconfig.json
@@ -18,12 +18,13 @@
     "isolatedModules": true,
     "jsx": "preserve",
     "jsxImportSource": "theme-ui",
-    "incremental": true
+    "incremental": true,
+    "baseUrl":"src"
   },
   "include": [
     "next-env.d.ts",
-    "**/*.ts",
-    "**/*.tsx"
+    "src/**/*.ts",
+    "src/**/*.tsx"
   ],
   "exclude": [
     "node_modules"
